-add element to vault now respond with the id of the new element -added a check that blocks the query if there are no token in existingTokens array

+++ src/utils/token_handler.js
@@ -1,8 +1,8 @@
 import jwt from 'jsonwebtoken';
 import config from "./config.js";
 import {userTokensCollection} from "./database.js";
-const {token} = config
 
+const {token} = config
 
 
 export function generateAccessToken(email, name, surname) {
@@ -31,17 +31,22 @@ export async function removeExistingRefreshToken(req, res, next) {
         let refreshTokens = await userTokensCollection.find({}).toArray().then(result => {
             return result
         })
-        let existingTokens= []
-        for(let obj of refreshTokens) {
+        let existingTokens = []
+        for (let obj of refreshTokens) {
             let payload = jwt.decode(obj.refreshToken, {complete: true}).payload
             if (payload.email === req.body.email) {
                 existingTokens.push(obj.refreshToken)
             }
         }
-        userTokensCollection.deleteMany({refreshToken: {$in: existingTokens}}).then(() => {
-            console.log(existingTokens)
-            console.log("Old Refresh Token removed from db")
-        })
+        if (existingTokens.length === 0) {
+            console.log("No old refresh token found")
+
+        } else {
+            userTokensCollection.deleteMany({refreshToken: {$in: existingTokens}}).then(() => {
+                console.log(existingTokens)
+                console.log("Old Refresh Token removed from db")
+            })
+        }
         next()
     } catch (e) {
         console.log(e)
